Remove unnecessary try-catch
Let parser catch the exception.
